+++ vitrina/structure/tasks.py
@@ -2,12 +2,13 @@ import contextlib
 import io
 
 from celery import shared_task
-from .models import ManifestValidationEntry, ValidationStatus
+from .models import ManifestValidationEntry, ValidationStatus, UMLDiagram, UMLDiagramStatus
 from spinta.core.context import create_context
 from spinta.cli.config import check
 from spinta.core.enums import Mode
 from types import SimpleNamespace
 from uuid import UUID
+from vitrina.structure.services import generate_mermaid_diagram
 
 
 @shared_task
@@ -31,3 +32,28 @@ def validate_manifest_task(manifest_id: UUID) -> None:
         manifest.error_message = logs
 
     manifest.save(update_fields=["validation_status", "error_message", "updated_at"])
+
+
+@shared_task
+def update_uml_diagram(uml_id: UUID) -> None:
+    uml_diagram = UMLDiagram.objects.prefetch_related("metadata_version__dataset").get(pk=uml_id)
+    current_version = uml_diagram.version_counter
+    try:
+        mermaid = generate_mermaid_diagram(uml_diagram.metadata_version.dataset, uml_diagram.metadata_version)
+    except Exception as e:
+        uml_diagram.refresh_from_db()
+        uml_diagram.status = (
+            UMLDiagramStatus.FAILED if uml_diagram.version_counter == current_version else UMLDiagramStatus.OUTDATED
+        )
+        uml_diagram.error_message = str(e) if uml_diagram.version_counter == current_version else None
+        uml_diagram.save(update_fields=["status", "error_message", "updated_at"])
+        return
+
+    uml_diagram.refresh_from_db()
+    if uml_diagram.version_counter == current_version:
+        uml_diagram.mermaid = mermaid
+        uml_diagram.status = UMLDiagramStatus.UP_TO_DATE
+    else:
+        uml_diagram.status = UMLDiagramStatus.OUTDATED
+    uml_diagram.error_message = None
+    uml_diagram.save(update_fields=["mermaid", "status", "error_message", "updated_at"])